refactor(mobile): move store settings and store into domain folder
