+++ mobile/lib/infrastructure/repositories/store.repository.dart
@@ -1,5 +1,5 @@
 import 'package:immich_mobile/domain/interfaces/store.interface.dart';
-import 'package:immich_mobile/entities/store.entity.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/user.entity.dart';
 import 'package:immich_mobile/infrastructure/entities/store.entity.dart';
 import 'package:immich_mobile/infrastructure/repositories/db.repository.dart';
