+++ mobile/lib/pages/share_intent/share_intent.page.dart
@@ -1,14 +1,14 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:flutter_hooks/flutter_hooks.dart' hide Store;
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
+import 'package:immich_mobile/entities/store.entity.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/models/upload/share_intent_attachment.model.dart';
-
 import 'package:immich_mobile/pages/common/large_leading_tile.dart';
 import 'package:immich_mobile/providers/asset_viewer/share_intent_upload.provider.dart';
-import 'package:immich_mobile/entities/store.entity.dart' as db_store;
 
 @RoutePage()
 class ShareIntentPage extends HookConsumerWidget {
@@ -18,8 +18,7 @@ class ShareIntentPage extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final currentEndpoint =
-        db_store.Store.get(db_store.StoreKey.serverEndpoint);
+    final currentEndpoint = Store.get(StoreKey.serverEndpoint);
     final candidates = ref.watch(shareIntentUploadProvider);
     final isUploaded = useState(false);
