+++ mobile/lib/services/app_settings.service.dart
@@ -1,4 +1,5 @@
 import 'package:immich_mobile/constants/colors.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/store.entity.dart';
 
 enum AppSettingsEnum<T> {
