+++ mobile/lib/services/device.service.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_udid/flutter_udid.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/store.entity.dart';
 
 final deviceServiceProvider = Provider((ref) => DeviceService());
