+++ mobile/lib/utils/http_ssl_cert_override.dart
@@ -1,6 +1,8 @@
 import 'dart:io';
-import 'package:immich_mobile/services/app_settings.service.dart';
+
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/store.entity.dart';
+import 'package:immich_mobile/services/app_settings.service.dart';
 import 'package:logging/logging.dart';
 
 class HttpSSLCertOverride extends HttpOverrides {
