+++ mobile/lib/utils/migration.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/album.entity.dart';
 import 'package:immich_mobile/entities/asset.entity.dart';
 import 'package:immich_mobile/entities/etag.entity.dart';
