+++ mobile/lib/utils/url_helper.dart
@@ -1,3 +1,4 @@
+import 'package:immich_mobile/domain/models/store.model.dart';
 import 'package:immich_mobile/entities/store.entity.dart';
 
 String sanitizeUrl(String url) {
