+++ mobile/lib/widgets/search/curated_places_row.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
+import 'package:immich_mobile/entities/store.entity.dart';
 import 'package:immich_mobile/models/search/search_curated_content.model.dart';
 import 'package:immich_mobile/widgets/search/search_map_thumbnail.dart';
 import 'package:immich_mobile/widgets/search/thumbnail_with_info.dart';
-import 'package:immich_mobile/entities/store.entity.dart';
 
 class CuratedPlacesRow extends StatelessWidget {
   const CuratedPlacesRow({
