+++ mobile/lib/widgets/settings/networking_settings/networking_settings.dart
@@ -1,7 +1,10 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:flutter_hooks/flutter_hooks.dart' hide Store;
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
+import 'package:immich_mobile/entities/store.entity.dart';
+import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/models/auth/auxilary_endpoint.model.dart';
 import 'package:immich_mobile/providers/network.provider.dart';
 import 'package:immich_mobile/services/app_settings.service.dart';
@@ -10,16 +13,12 @@ import 'package:immich_mobile/widgets/settings/networking_settings/external_netw
 import 'package:immich_mobile/widgets/settings/networking_settings/local_network_preference.dart';
 import 'package:immich_mobile/widgets/settings/settings_switch_list_tile.dart';
 
-import 'package:immich_mobile/entities/store.entity.dart' as db_store;
-import 'package:immich_mobile/extensions/build_context_extensions.dart';
-
 class NetworkingSettings extends HookConsumerWidget {
   const NetworkingSettings({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final currentEndpoint =
-        db_store.Store.get(db_store.StoreKey.serverEndpoint);
+    final currentEndpoint = Store.get(StoreKey.serverEndpoint);
     final featureEnabled =
         useAppSettingsState(AppSettingsEnum.autoEndpointSwitching);
